pypet2bids/pyproject.toml
[tool.poetry]
name = "pypet2bids"
version = "0.0.3"
description = "A python implementation of an ECAT to BIDS converter."
authors = ["anthony galassi <28850131+bendhouseart@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">3.7.1,<3.10"
nibabel = "^3.2.1"
numpy = "^1.21.3"
pyparsing = "^3.0.4"
python-dateutil = "^2.8.2"
python-dotenv = "^0.19.1"
scipy = "^1.7.1"
six = "^1.16.0"
pytest = "^6.2.5"
pydicom = "^2.2.2"
Gooey = "^1.0.8"
pandas = "^1.3.4"
openpyxl = "^3.0.9"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
pypet2bids = 'pypet2bids.cli:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
